+++ src/fetch-jsonp.js
@@ -1,5 +1,5 @@
 const defaultOptions = {
-  timeout: 5000,
+  timeout: 0,
   jsonpCallback: 'callback',
   jsonpCallbackFunction: null,
 };
@@ -9,13 +9,13 @@ function generateCallbackFunction() {
 }
 
 // Known issue: Will throw 'Uncaught ReferenceError: callback_*** is not defined' error if request timeout
-function clearFunction(functionName) {
+function clearGlobalObject(objectName) {
   // IE8 throws an exception when you try to delete a property on window
   // http://stackoverflow.com/a/1824228/751089
   try {
-    delete window[functionName];
-  } catch(e) {
-    window[functionName] = undefined;
+    delete window[objectName];
+  } catch (e) {
+    window[objectName] = undefined;
   }
 }
 
@@ -27,11 +27,21 @@ function removeScript(scriptId) {
 const fetchJsonp = function(url, options = {}) {
   const timeout = options.timeout != null ? options.timeout : defaultOptions.timeout;
   const jsonpCallback = options.jsonpCallback != null ? options.jsonpCallback : defaultOptions.jsonpCallback;
+  //Generate different "timeoutid"
 
-  let timeoutId;
+  //Do not use the same timeoutid
+  //let timeoutId;
 
   return new Promise((resolve, reject) => {
     let callbackFunction = options.jsonpCallbackFunction || generateCallbackFunction();
+
+    const timeoutId = jsonpCallback + '_' + callbackFunction + '_' + 'timeoutId';
+
+    function claer() {
+      clearGlobalObject(timeoutId);
+      clearGlobalObject(callbackFunction);
+      removeScript(jsonpCallback + '_' + callbackFunction);
+    }
 
     window[callbackFunction] = function(response) {
       resolve({
@@ -44,9 +54,7 @@ const fetchJsonp = function(url, options = {}) {
 
       if (timeoutId) clearTimeout(timeoutId);
 
-      removeScript(jsonpCallback + '_' + callbackFunction);
-
-      clearFunction(callbackFunction);
+      claer();
     };
 
     // Check if the user set their own params, and if not add a ? to start a list of params
@@ -55,14 +63,37 @@ const fetchJsonp = function(url, options = {}) {
     const jsonpScript = document.createElement('script');
     jsonpScript.setAttribute("src", url + jsonpCallback + '=' + callbackFunction);
     jsonpScript.id = jsonpCallback + '_' + callbackFunction;
+
+    // Attach handlers for all browsers
+    jsonpScript.onload = jsonpScript.onreadystatechange = function(err) {
+      var hasError = false;
+      //IE8 Unable to determine the load failed.
+      //But the execution is faster than the this.readyState change.
+      //Request to complete the callback will be deleted, if there is a callback, that is, the request failed
+      //
+      //Delay will affect the result but does not affect your overall request
+      // status, which is considered a failure by more than your delay time.
+      if (/loaded|complete/.test(this.readyState)) {
+        hasError = !!window[callbackFunction];
+      }
+
+      if (hasError || err) {
+        reject(new Error(`JSONP request to ${url} request failed`));
+        claer();
+      }
+    }
+
+    if (timeout > 0) {
+      //Generate different “timeoutid”
+      //That is not because of the asynchronous lead, the entire "timeout" chaos
+      window[timeoutid] = setTimeout(() => {
+        reject(new Error(`JSONP request to ${url} timed out`));
+        claer();
+      }, timeout);
+    }
+
     document.getElementsByTagName("head")[0].appendChild(jsonpScript);
 
-    timeoutId = setTimeout(() => {
-      reject(new Error(`JSONP request to ${url} timed out`));
-
-      clearFunction(callbackFunction);
-      removeScript(jsonpCallback + '_' + callbackFunction);
-    }, timeout);
   });
 };