+++ HISTORY.md
@@ -1,3 +1,8 @@
+1.1.0 / 2016-1-29
+==================
+* Add IE8 onerror and onreadystatechange check error
+* Each request to produce a different "timeoutid"
+
 1.0.0 / 2015-11-19
 ==================
 * Remove Bower support